Update manual.yml
Proposing a fix to the jq manual's section on the `-e`/ `--exit-status` flag. As of version 1.6, running `jq` and getting a parse error results in exit code 4. I don't know if this is something wrong with the manual or the actual code itself, but I figured I'd bring it to the project's attention this way.